+++ packages/react-liveness/src/components/FaceLivenessDetector/service/types/machine.ts
@@ -8,7 +8,7 @@ import {
 } from '@aws-sdk/client-rekognitionstreaming';
 
 import {
-  FaceLivenessDetectorProps,
+  FaceLivenessDetectorCoreProps,
   FaceMatchState,
   LivenessOvalDetails,
   IlluminationState,
@@ -60,7 +60,7 @@ export type LivenessContext = Partial<HydratedLivenessContext>;
 export interface HydratedLivenessContext {
   maxFailedAttempts: number;
   failedAttempts: number;
-  componentProps: FaceLivenessDetectorProps;
+  componentProps: FaceLivenessDetectorCoreProps;
   serverSessionInformation: SessionInformation;
   challengeId: string;
   videoAssociatedParams: VideoAssociatedParams;