+++ packages/react-liveness/src/components/FaceLivenessDetector/service/utils/__tests__/streamProvider.test.ts
@@ -7,6 +7,7 @@ import { RekognitionStreamingClient } from '@aws-sdk/client-rekognitionstreaming
 import { LivenessStreamProvider } from '../streamProvider';
 import { VideoRecorder } from '../videoRecorder';
 import { mockClientSessionInformationEvent } from '../__mocks__/testUtils';
+import { AwsCredentialProvider } from '../../types';
 
 jest.mock('../videoRecorder');
 jest.mock('@aws-sdk/client-rekognitionstreaming');
@@ -111,35 +112,53 @@ describe('LivenessStreamProvider', () => {
 
   describe('constructor', () => {
     test('happy case', () => {
-      new LivenessStreamProvider(
-        'sessionId',
-        'us-east-1',
-        mockVideoMediaStream,
-        mockVideoEl
-      );
+      new LivenessStreamProvider({
+        sessionId: 'sessionId',
+        region: 'us-east-1',
+        stream: mockVideoMediaStream,
+        videoEl: mockVideoEl,
+      });
+    });
+
+    test('with credential provider', () => {
+      const credentialProvider: AwsCredentialProvider = async () => {
+        return {
+          accessKeyId: 'test',
+          secretAccessKey: 'test',
+          sessionToken: 'test',
+          expiration: new Date(),
+        };
+      };
+      new LivenessStreamProvider({
+        sessionId: 'sessionId',
+        region: 'us-east-1',
+        stream: mockVideoMediaStream,
+        videoEl: mockVideoEl,
+        credentialProvider,
+      });
     });
   });
 
   describe('getResponseStream', () => {
     test('happy case', async () => {
-      const provider = new LivenessStreamProvider(
-        'sessionId',
-        'us-east-1',
-        mockVideoMediaStream,
-        mockVideoEl
-      );
+      const provider = new LivenessStreamProvider({
+        sessionId: 'sessionId',
+        region: 'us-east-1',
+        stream: mockVideoMediaStream,
+        videoEl: mockVideoEl,
+      });
       expect(await provider.getResponseStream()).toBeDefined();
     });
   });
 
   describe('startLivenessVideoConnection', () => {
     test('happy case', async () => {
-      const provider = new LivenessStreamProvider(
-        'sessionId',
-        'us-east-1',
-        mockVideoMediaStream,
-        mockVideoEl
-      );
+      const provider = new LivenessStreamProvider({
+        sessionId: 'sessionId',
+        region: 'us-east-1',
+        stream: mockVideoMediaStream,
+        videoEl: mockVideoEl,
+      });
       await provider.startRecordingLivenessVideo();
       expect(mockVideoRecorder.start).toHaveBeenCalledTimes(1);
     });
@@ -147,12 +166,12 @@ describe('LivenessStreamProvider', () => {
 
   describe('getAsyncGeneratorFromReadableStream', () => {
     test('yield video chunk events', async () => {
-      const provider = new LivenessStreamProvider(
-        'sessionId',
-        'us-east-1',
-        mockVideoMediaStream,
-        mockVideoEl
-      );
+      const provider = new LivenessStreamProvider({
+        sessionId: 'sessionId',
+        region: 'us-east-1',
+        stream: mockVideoMediaStream,
+        videoEl: mockVideoEl,
+      });
       const requestStream = (
         provider as any
       ).getAsyncGeneratorFromReadableStream(mockReadableStream)();
@@ -164,12 +183,12 @@ describe('LivenessStreamProvider', () => {
     });
 
     test('does not yield empty video chunks', async () => {
-      const provider = new LivenessStreamProvider(
-        'sessionId',
-        'us-east-1',
-        mockVideoMediaStream,
-        mockVideoEl
-      );
+      const provider = new LivenessStreamProvider({
+        sessionId: 'sessionId',
+        region: 'us-east-1',
+        stream: mockVideoMediaStream,
+        videoEl: mockVideoEl,
+      });
       const requestStream = (
         provider as any
       ).getAsyncGeneratorFromReadableStream(
@@ -185,12 +204,12 @@ describe('LivenessStreamProvider', () => {
 
   describe('sendClientInfo', () => {
     test('happy case', async () => {
-      const provider = new LivenessStreamProvider(
-        'sessionId',
-        'us-east-1',
-        mockVideoMediaStream,
-        mockVideoEl
-      );
+      const provider = new LivenessStreamProvider({
+        sessionId: 'sessionId',
+        region: 'us-east-1',
+        stream: mockVideoMediaStream,
+        videoEl: mockVideoEl,
+      });
       await provider.sendClientInfo(mockClientSessionInformationEvent);
 
       expect(mockVideoRecorder.dispatch).toHaveBeenCalledTimes(1);
@@ -199,12 +218,12 @@ describe('LivenessStreamProvider', () => {
 
   describe('stopVideo', () => {
     test('should stop sending video events', async () => {
-      const provider = new LivenessStreamProvider(
-        'sessionId',
-        'us-east-1',
-        mockVideoMediaStream,
-        mockVideoEl
-      );
+      const provider = new LivenessStreamProvider({
+        sessionId: 'sessionId',
+        region: 'us-east-1',
+        stream: mockVideoMediaStream,
+        videoEl: mockVideoEl,
+      });
       const response = await provider.stopVideo();
       expect(mockVideoRecorder.stop).toHaveBeenCalled();
       expect(response).toBeUndefined();
@@ -213,12 +232,12 @@ describe('LivenessStreamProvider', () => {
 
   describe('dispatchStopVideoEvent', () => {
     test('should dispatch an empty video chunk', async () => {
-      const provider = new LivenessStreamProvider(
-        'sessionId',
-        'us-east-1',
-        mockVideoMediaStream,
-        mockVideoEl
-      );
+      const provider = new LivenessStreamProvider({
+        sessionId: 'sessionId',
+        region: 'us-east-1',
+        stream: mockVideoMediaStream,
+        videoEl: mockVideoEl,
+      });
       const response = await provider.dispatchStopVideoEvent();
       expect(mockVideoRecorder.dispatch).toHaveBeenCalledTimes(2);
       expect(response).toBeUndefined();
@@ -227,12 +246,12 @@ describe('LivenessStreamProvider', () => {
 
   describe('endStream', () => {
     test('should stop video and end the stream and return a promise if cancelled successfully', async () => {
-      const provider = new LivenessStreamProvider(
-        'sessionId',
-        'us-east-1',
-        mockVideoMediaStream,
-        mockVideoEl
-      );
+      const provider = new LivenessStreamProvider({
+        sessionId: 'sessionId',
+        region: 'us-east-1',
+        stream: mockVideoMediaStream,
+        videoEl: mockVideoEl,
+      });
       const response = await provider.endStream();
 
       expect(mockVideoRecorder.stop).toHaveBeenCalled();
@@ -240,12 +259,12 @@ describe('LivenessStreamProvider', () => {
     });
 
     test('should stop video even if the stream is not available', async () => {
-      const provider = new LivenessStreamProvider(
-        'sessionId',
-        'us-east-1',
-        mockVideoMediaStream,
-        mockVideoEl
-      );
+      const provider = new LivenessStreamProvider({
+        sessionId: 'sessionId',
+        region: 'us-east-1',
+        stream: mockVideoMediaStream,
+        videoEl: mockVideoEl,
+      });
       (provider as any)._reader = undefined;
       const response = await provider.endStream();