+++ packages/ui/src/helpers/authenticator/__tests__/utils.test.ts
@@ -1,4 +1,9 @@
-import { configureComponent, getTotpCodeURL, trimValues } from '../utils';
+import {
+  configureComponent,
+  getTotpCodeURL,
+  isValidEmail,
+  trimValues,
+} from '../utils';
 
 import * as AuthModule from '@aws-amplify/auth';
 
@@ -72,5 +77,26 @@ describe('configureComponent', () => {
     expect(appendToCognitoUserAgentSpy).toHaveBeenCalledWith(
       `${packageName}/${version}`
     );
+  });
+});
+
+describe('isValidEmail', () => {
+  it('should return true for a valid email address', () => {
+    expect(isValidEmail('test@example.com')).toBe(true);
+    expect(isValidEmail('TEST@EXAMPLE.COM')).toBe(true);
+  });
+
+  it('should return false for an invalid email address', () => {
+    expect(isValidEmail('testexample.com')).toBe(false);
+    expect(isValidEmail('test@')).toBe(false);
+    expect(isValidEmail('test@.')).toBe(false);
+    expect(isValidEmail('test@example@test.com')).toBe(false);
+    expect(isValidEmail('test @example.com')).toBe(false);
+  });
+
+  it('should return false if there is no email address', () => {
+    expect(isValidEmail(null)).toBe(false);
+    expect(isValidEmail(undefined)).toBe(false);
+    expect(isValidEmail('')).toBe(false);
   });
 });